# api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Blogging para Professores",
    "description": "API REST para criação e gerenciamento de posts educacionais",
    "version": "1.0.0",
    "contact": {
      "name": "Tech Challenge Team",
      "email": "team@techchallenge.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de Desenvolvimento"
    }
  ],
  "paths": {
    "/posts": {
      "get": {
        "summary": "Listar Posts",
        "description": "Retorna lista de posts publicados com paginação",
        "tags": ["Posts"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Número da página",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Posts por página",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Ordenação",
            "schema": {
              "type": "string",
              "default": "-createdAt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Posts recuperados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "posts": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Post"
                          }
                        },
                        "totalPosts": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        },
                        "hasNextPage": {
                          "type": "boolean"
                        },
                        "hasPrevPage": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar Post",
        "description": "Cria um novo post (apenas para professores)",
        "tags": ["Posts"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "content", "author"],
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 200,
                    "example": "Título do Post"
                  },
                  "content": {
                    "type": "string",
                    "minLength": 10,
                    "example": "Conteúdo do post..."
                  },
                  "author": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 100,
                    "example": "Nome do Professor"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["tag1", "tag2"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Post"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Post": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID único do post"
          },
          "title": {
            "type": "string",
            "description": "Título do post"
          },
          "content": {
            "type": "string",
            "description": "Conteúdo do post"
          },
          "author": {
            "type": "string",
            "description": "Autor do post"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Tags do post"
          },
          "isPublished": {
            "type": "boolean",
            "description": "Status de publicação"
          },
          "readTime": {
            "type": "integer",
            "description": "Tempo estimado de leitura em minutos"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de última atualização"
          }
        }
      }
    }
  }
} 